Move Canvas and Image class into their own files
